Refactor settings screen to Jatpack Compose
