+++ app/src/main/java/com/example/playlistmaker/ui/media/compose/PlaylistGrid.kt
@@ -1,0 +1,27 @@
+package com.example.playlistmaker.ui.media.compose
+
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.lazy.grid.itemsIndexed
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.unit.dp
+import com.example.playlistmaker.domain.mediateca.playlists.model.Playlist
+
+@Composable
+fun PlaylistGrid(
+    playlists: List<Playlist>,
+    onPlaylistClick: (Playlist) -> Unit,
+) {
+    LazyVerticalGrid(
+        columns = GridCells.Fixed(2),
+        contentPadding = PaddingValues(8.dp)
+    ) {
+        itemsIndexed(playlists) { _, playlist ->
+            PlaylistItem(
+                playlist = playlist,
+                onClick = { onPlaylistClick(playlist) },
+            )
+        }
+    }
+}
