+++ app/src/main/java/com/example/playlistmaker/ui/settings/compose/SettingItem.kt
@@ -1,0 +1,54 @@
+package com.example.playlistmaker.ui.settings.compose
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Icon
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.Font
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.playlistmaker.R
+
+@Composable
+fun SettingItem(
+    title: String,
+    icon: ImageVector,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Row(
+        modifier = modifier
+            .fillMaxWidth()
+            .clickable(onClick = onClick)
+            .padding(16.dp),
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.SpaceBetween
+    ) {
+        Text(
+            text = title,
+            style = TextStyle(
+                color = colorResource(R.color.black_day_white_night),
+                fontSize = 16.sp,
+                fontFamily = FontFamily(
+                    Font(R.font.ys_display_regular, weight = FontWeight.W400)
+                )
+            ),
+        )
+        Icon(
+            imageVector = icon,
+            contentDescription = title,
+            tint = colorResource(R.color.grey2_day_true_white_night)
+        )
+    }
+}
