+++ app/src/main/java/com/example/playlistmaker/ui/settings/compose/ThemeSwitcher.kt
@@ -1,0 +1,49 @@
+package com.example.playlistmaker.ui.settings.compose
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Switch
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.Font
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.playlistmaker.R
+
+@Composable
+fun ThemeSwitcher(
+    isDarkTheme: Boolean,
+    onThemeChanged: (Boolean) -> Unit,
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(start = 16.dp, end = 16.dp),
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.SpaceBetween
+    ) {
+        Text(
+            text = stringResource(R.string.theme),
+            style = TextStyle(
+                color = colorResource(R.color.black_day_white_night),
+                fontSize = 16.sp,
+                fontFamily = FontFamily(
+                    Font(R.font.ys_display_regular, weight = FontWeight.W400)
+                )
+            ),
+        )
+        Switch(
+            checked = isDarkTheme,
+            onCheckedChange = onThemeChanged
+        )
+    }
+}
