+++ app/src/main/res/navigation/navigation_graph.xml
@@ -16,7 +16,7 @@
     </fragment>
     <fragment
         android:id="@+id/settingsFragment"
-        android:name="com.example.playlistmaker.ui.settings.fragment.SettingsFragment"
+        android:name="com.example.playlistmaker.ui.settings.fragment.SettingsFragmentWithComposeUI"
         android:label="fragment_settings"
         tools:layout="@layout/fragment_settings" />
     <fragment
